Move DataTemplateSelector to own file in template
